% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_colour_YlGrBl_d}
\alias{scale_colour_YlGrBl_d}
\title{YlGrBl discrete scales color from YlGrBlPal}
\usage{
scale_colour_YlGrBl_d(direction = 1)
}
\arguments{
\item{direction}{To change the direction of the return when direction < 1}
}
\description{
YlGrBl discrete scales color from YlGrBlPal
}
\examples{
{library('ggplot2')
ggplot(data = palmerpenguins::penguins, aes(x = flipper_length_mm, y = body_mass_g)) +
geom_point(aes(color = species))+
scale_colour_YlGrBl_d()}
}
