% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_YlGrBl_seq_c}
\alias{scale_colour_YlGrBl_seq_c}
\title{YlGrBl continuous scales color from YlGrBlPal}
\usage{
scale_colour_YlGrBl_seq_c(name, direction = 1, amount = 1)
}
\arguments{
\item{name}{Name of the 5 colors including yellow, green, turquoise, blue, indigo.}

\item{direction}{Reverse the luminance of the color at top of scale}

\item{amount}{To define the amount of luminance}
}
\description{
YlGrBl continuous scales color from YlGrBlPal
}
\examples{
{library('ggplot2')
ggplot(data = palmerpenguins::penguins, aes(x = flipper_length_mm, y = body_mass_g)) +
geom_point(aes(color = body_mass_g))+
scale_colour_YlGrBl_seq_c(name = 'blue', direction = 1, amount = 1)}
}
